+++ test/integration_tests/MacrosTest.cpp
@@ -35,9 +35,9 @@ TEST_CASE("macros")
 
   Logger* logger = Frontend::create_or_get_logger(
     logger_name, std::move(file_sink),
-    PatternFormatterOptions{
-      "%(time) [%(thread_id)] %(short_source_location:<28) LOG_%(log_level:<9) "
-      "%(logger:<12) %(message) [ %(tags)] [%(named_args)]"});
+    PatternFormatterOptions{"%(time) [%(thread_id)] %(short_source_location:<28) "
+                            "%(log_level_short_code) LOG_%(log_level:<9) "
+                            "%(logger:<12) %(message) [ %(tags)] [%(named_args)]"});
   logger->init_backtrace(1, quill::LogLevel::Error);
   logger->set_log_level(quill::LogLevel::TraceL3);
 
@@ -148,152 +148,172 @@ TEST_CASE("macros")
 
   // Read file and check
   std::vector<std::string> const file_contents = quill::testing::file_contents(filename);
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L3  logger       L3: 1"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L2  logger       L2: 2"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L1  logger       L1: 3"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_DEBUG     logger       DBG: 4"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_INFO      logger       INF: 5"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_NOTICE    logger       NTC: N"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_WARNING   logger       WRN: W"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_ERROR     logger       ERR: 6.78"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       CRT: -1"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       DYN: 42"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_BACKTRACE logger       BT: 255"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T3 LOG_TRACE_L3  logger       L3: 1"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T2 LOG_TRACE_L2  logger       L2: 2"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T1 LOG_TRACE_L1  logger       L1: 3"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"D LOG_DEBUG     logger       DBG: 4"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"I LOG_INFO      logger       INF: 5"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"N LOG_NOTICE    logger       NTC: N"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"W LOG_WARNING   logger       WRN: W"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"E LOG_ERROR     logger       ERR: 6.78"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"C LOG_CRITICAL  logger       CRT: -1"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"C LOG_CRITICAL  logger       DYN: 42"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"BT LOG_BACKTRACE logger       BT: 255"}));
 
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L3  logger       B L3: 10"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L2  logger       B L2: 20"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L1  logger       B L1: 30"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_DEBUG     logger       B DBG: 40"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_INFO      logger       B INF: 50"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_NOTICE    logger       B NTC: Notice"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_WARNING   logger       B WRN: Warning"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_ERROR     logger       B ERR: 60.78"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       B CRT: -10"}));
-
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L3  logger       C L3: 100 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L2  logger       C L2: 200 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L1  logger       C L1: 300 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_DEBUG     logger       C DBG: 400 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_INFO      logger       C INF: 500 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_NOTICE    logger       C NTC: NoticeTag [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_WARNING   logger       C WRN: WarningTag [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_ERROR     logger       C ERR: 600.78 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       C CRT: -100 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       C DYN: 420 [ #tag ]"}));
-  REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_BACKTRACE logger       C BT: 255 [ #tag ]"}));
-
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L3  logger       D L3 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L2  logger       D L2 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L1  logger       D L1 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_DEBUG     logger       D DBG [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_INFO      logger       D INF [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_NOTICE    logger       D NTC [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_WARNING   logger       D WRN [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_ERROR     logger       D ERR [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       D CRT [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       D DYN [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_BACKTRACE logger       D BT [var: 1337]"}));
-
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L3  logger       E L3 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L2  logger       E L2 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_TRACE_L1  logger       E L1 [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_DEBUG     logger       E DBG [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_INFO      logger       E INF [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_NOTICE    logger       E NTC [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_WARNING   logger       E WRN [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_ERROR     logger       E ERR [var: 1337]"}));
-  REQUIRE(quill::testing::file_contains(file_contents, std::string{"LOG_CRITICAL  logger       E CRT [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T3 LOG_TRACE_L3  logger       B L3: 10"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T2 LOG_TRACE_L2  logger       B L2: 20"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"T1 LOG_TRACE_L1  logger       B L1: 30"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"D LOG_DEBUG     logger       B DBG: 40"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"I LOG_INFO      logger       B INF: 50"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"N LOG_NOTICE    logger       B NTC: Notice"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"W LOG_WARNING   logger       B WRN: Warning"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"E LOG_ERROR     logger       B ERR: 60.78"}));
+  REQUIRE(quill::testing::file_contains(file_contents, std::string{"C LOG_CRITICAL  logger       B CRT: -10"}));
 
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L3  logger       F L3 [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       C L3: 100 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L2  logger       F L2 [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       C L2: 200 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L1  logger       F L1 [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       C L1: 300 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_DEBUG     logger       F DBG [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"D LOG_DEBUG     logger       C DBG: 400 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_INFO      logger       F INF [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"I LOG_INFO      logger       C INF: 500 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_NOTICE    logger       F NTC [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"N LOG_NOTICE    logger       C NTC: NoticeTag [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_WARNING   logger       F WRN [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"W LOG_WARNING   logger       C WRN: WarningTag [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_ERROR     logger       F ERR [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"E LOG_ERROR     logger       C ERR: 600.78 [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       F CRT [var: 1337] [ #tag ]"}));
+    file_contents, std::string{"C LOG_CRITICAL  logger       C CRT: -100 [ #tag ]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"C LOG_CRITICAL  logger       C DYN: 420 [ #tag ]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"BT LOG_BACKTRACE logger       C BT: 255 [ #tag ]"}));
 
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L3  logger       G L3 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       D L3 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L2  logger       G L2 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       D L2 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L1  logger       G L1 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       D L1 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_DEBUG     logger       G DBG 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"D LOG_DEBUG     logger       D DBG [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_INFO      logger       G INF 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"I LOG_INFO      logger       D INF [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_NOTICE    logger       G NTC 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"N LOG_NOTICE    logger       D NTC [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_WARNING   logger       G WRN 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"W LOG_WARNING   logger       D WRN [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_ERROR     logger       G ERR 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"E LOG_ERROR     logger       D ERR [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       G CRT 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"C LOG_CRITICAL  logger       D CRT [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       G DYN 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"C LOG_CRITICAL  logger       D DYN [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_BACKTRACE logger       G BT 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"BT LOG_BACKTRACE logger       D BT [var: 1337]"}));
 
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L3  logger       H L3 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       E L3 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L2  logger       H L2 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       E L2 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L1  logger       H L1 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       E L1 [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_DEBUG     logger       H DBG 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"D LOG_DEBUG     logger       E DBG [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_INFO      logger       H INF 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"I LOG_INFO      logger       E INF [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_NOTICE    logger       H NTC 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"N LOG_NOTICE    logger       E NTC [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_WARNING   logger       H WRN 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"W LOG_WARNING   logger       E WRN [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_ERROR     logger       H ERR 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"E LOG_ERROR     logger       E ERR [var: 1337]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       H CRT 1337 [ ] [var: 1337]"}));
+    file_contents, std::string{"C LOG_CRITICAL  logger       E CRT [var: 1337]"}));
 
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L3  logger       K L3 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       F L3 [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L2  logger       K L2 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       F L2 [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_TRACE_L1  logger       K L1 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       F L1 [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_DEBUG     logger       K DBG 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"D LOG_DEBUG     logger       F DBG [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_INFO      logger       K INF 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"I LOG_INFO      logger       F INF [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_NOTICE    logger       K NTC 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"N LOG_NOTICE    logger       F NTC [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_WARNING   logger       K WRN 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"W LOG_WARNING   logger       F WRN [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_ERROR     logger       K ERR 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"E LOG_ERROR     logger       F ERR [var: 1337] [ #tag ]"}));
   REQUIRE(quill::testing::file_contains(
-    file_contents, std::string{"LOG_CRITICAL  logger       K CRT 1337 [ #tag ] [var: 1337]"}));
+    file_contents, std::string{"C LOG_CRITICAL  logger       F CRT [var: 1337] [ #tag ]"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       G L3 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       G L2 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       G L1 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"D LOG_DEBUG     logger       G DBG 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"I LOG_INFO      logger       G INF 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"N LOG_NOTICE    logger       G NTC 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"W LOG_WARNING   logger       G WRN 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"E LOG_ERROR     logger       G ERR 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"C LOG_CRITICAL  logger       G CRT 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"C LOG_CRITICAL  logger       G DYN 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"BT LOG_BACKTRACE logger       G BT 1337 [ ] [var: 1337]"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       H L3 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       H L2 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       H L1 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"D LOG_DEBUG     logger       H DBG 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"I LOG_INFO      logger       H INF 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"N LOG_NOTICE    logger       H NTC 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"W LOG_WARNING   logger       H WRN 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"E LOG_ERROR     logger       H ERR 1337 [ ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"C LOG_CRITICAL  logger       H CRT 1337 [ ] [var: 1337]"}));
+
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T3 LOG_TRACE_L3  logger       K L3 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T2 LOG_TRACE_L2  logger       K L2 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"T1 LOG_TRACE_L1  logger       K L1 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"D LOG_DEBUG     logger       K DBG 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"I LOG_INFO      logger       K INF 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"N LOG_NOTICE    logger       K NTC 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"W LOG_WARNING   logger       K WRN 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"E LOG_ERROR     logger       K ERR 1337 [ #tag ] [var: 1337]"}));
+  REQUIRE(quill::testing::file_contains(
+    file_contents, std::string{"C LOG_CRITICAL  logger       K CRT 1337 [ #tag ] [var: 1337]"}));
 
   testing::remove_file(filename);
 }